requirements.txt
Django==4.2.7
django-taggit==4.0.0
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
